Moved Dom and Selector assertions from ActionDispatch to ActionView.
