+++ actionpack/lib/action_dispatch/testing/integration.rb
@@ -2,6 +2,7 @@ require 'stringio'
 require 'uri'
 require 'active_support/core_ext/kernel/singleton_class'
 require 'active_support/core_ext/object/try'
+require 'action_view/testing/assertions'
 require 'rack/test'
 
 module ActionDispatch
@@ -313,7 +314,7 @@ module ActionDispatch
     end
 
     module Runner
-      include ActionDispatch::Assertions
+      include ActionDispatch::Assertions, ActionView::Assertions
 
       def app
         @app ||= nil
